+++ src/config/ProviderConfiguration.java
@@ -57,8 +57,8 @@ public class ProviderConfiguration extends Configuration{
 		int numberOfPlans = Integer.valueOf(currentProperties.getProperty(NUM_OF_PROVIDERS));
 		providers = new HashMap<String, Provider>();
 		for(int i = 1; i <= numberOfPlans; i++){
-			providers.put(currentProperties.getProperty(PROVIDER+i+NAME), new Provider(currentProperties.getProperty(PROVIDER+i+NAME), Double.valueOf(currentProperties.getProperty(PROVIDER+i+CPU_COST)), Double.valueOf(currentProperties.getProperty(PROVIDER+i+ON_DEMAND_LIMIT)),
-					Double.valueOf(currentProperties.getProperty(PROVIDER+i+RESERVATION_LIMIT)), Double.valueOf(currentProperties.getProperty(PROVIDER+i+ONE_YEAR_FEE)), 
+			providers.put(currentProperties.getProperty(PROVIDER+i+NAME), new Provider(currentProperties.getProperty(PROVIDER+i+NAME), Double.valueOf(currentProperties.getProperty(PROVIDER+i+CPU_COST)), Integer.valueOf(currentProperties.getProperty(PROVIDER+i+ON_DEMAND_LIMIT)),
+					Integer.valueOf(currentProperties.getProperty(PROVIDER+i+RESERVATION_LIMIT)), Double.valueOf(currentProperties.getProperty(PROVIDER+i+ONE_YEAR_FEE)), 
 					Double.valueOf(currentProperties.getProperty(PROVIDER+i+THREE_YEARS_FEE)), Double.valueOf(currentProperties.getProperty(PROVIDER+i+MONITORING)),
 					currentProperties.getProperty(PROVIDER+i+TRANSFER_IN), currentProperties.getProperty(PROVIDER+i+COST_TRANSFER_IN), currentProperties.getProperty(PROVIDER+i+TRANSFER_OUT), 
 					currentProperties.getProperty(PROVIDER+i+COST_TRANSFER_OUT)
